refactor!: change the logic for date of completion of a habit, making it client driven
BREAKING CHANGE: receive date of completion as a date string from client, and save as abstract date in UTC as
a standard for normalization.
api call fails if no date is provided. A date string is mandatory.
finding an existing completed habit now is returned using .lean(), effectively reducing ram usage
and time for querying.